+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Video and giphy attachments now use `makeMessageViewModifier`
 - Updated scalling of avatar images
 
+### 🐞 Fixed
+- `AttachmentTextView` respects configured body font
+
 # [4.34.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.34.0)
 _July 06, 2023_