+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### 🔄 Changed
+
+# [4.35.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.35.0)
+_August 09, 2023_
+
+### 🔄 Changed
 - Video and giphy attachments now use `makeMessageViewModifier`
 - Updated scalling of avatar images
 - Turn off channel updates when message thread shown